pyproject.toml
[tool.poetry]
name = "stocker"
version = "0.1.0"
description = ""
authors = ["pkwarc <piotr.j.kwarcinski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.7"
celery = "^5.1.2"
requests = "^2.26.0"
yfinance = "^0.1.63"
pytest-django = "^4.4.0"
psycopg2-binary = "^2.9.1"
djangorestframework = "^3.12.4"
black = "^21.9b0"
coreapi = "^2.3.3"
gunicorn = "^20.1.0"
djangorestframework-simplejwt = "^5.0.0"
redis = "^4.0.2"
django-health-check = "^3.16.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
